[us14] Generify findOrCreateBaseFolder in fileService
